+++ internal/servers/worker/handler.go
@@ -153,12 +153,7 @@ func (w *Worker) handleProxy() http.HandlerFunc {
 		}
 
 		defer func() {
-			connectionId := ci.id
-			if err := w.closeConnections(r.Context(), map[string]string{
-				connectionId: si.id,
-			}); err != nil {
-				w.logger.Error("error marking connection closed", "error", err, "connection_id", connectionId)
-			}
+			w.closeConnections(r.Context(), map[string]string{ci.id: si.id})
 		}()
 
 		si.Lock()