+++ internal/servers/worker/session.go
@@ -16,7 +16,8 @@ import (
 )
 
 const (
-	validateSessionTimeout = 90 * time.Second
+	validateSessionTimeout              = 90 * time.Second
+	errMakeSessionCloseInfoNilCloseInfo = "nil closeInfo supplied to makeSessionCloseInfo, this is a bug, please report it"
 )
 
 type connInfo struct {
@@ -251,49 +252,132 @@ func (w *Worker) closeConnection(ctx context.Context, req *pbs.CloseConnectionRe
 	return resp, nil
 }
 
-func (w *Worker) closeConnections(ctx context.Context, closeMap map[string]string) error {
-	w.logger.Trace("marking connections as closed", "session_and_connection_ids", fmt.Sprintf("%#v", closeMap))
+// closeConnections is a helper worker function that sends connection
+// close requests to the controller, and sets close times within the
+// worker. It is called during the worker status loop and on
+// connection exit on the proxy.
+//
+// closeInfo is a map of connections mapped to their individual
+// session.
+func (w *Worker) closeConnections(ctx context.Context, closeInfo map[string]string) {
+	if closeInfo == nil {
+		// This should not happen, but it's a no-op if it does. Just
+		// return.
+		return
+	}
 
-	closeData := make([]*pbs.CloseConnectionRequestData, 0, len(closeMap))
-	for connId := range closeMap {
+	w.logger.Trace("marking connections as closed", "session_and_connection_ids", fmt.Sprintf("%#v", closeInfo))
+	response, err := w.closeConnection(ctx, w.makeCloseConnectionRequest(closeInfo))
+	if err != nil {
+		w.logger.Error("error marking connections closed", "error", err)
+		w.logger.Warn(
+			"error contacting controller, connections will be closed only on worker",
+			"session_and_connection_ids", fmt.Sprintf("%#v", closeInfo),
+		)
+	}
+
+	// Mark connections as closed
+	closedIds, errs := w.setCloseTimeForResponse(w.makeSessionCloseInfo(closeInfo, response))
+	if len(errs) > 0 {
+		for _, err := range errs {
+			w.logger.Error("error marking connection closed in state", "err", err)
+		}
+	}
+
+	w.logger.Trace("connections successfully marked closed", "connection_ids", closedIds)
+}
+
+// makeCloseConnectionRequest creates a CloseConnectionRequest for
+// use with closing connections.
+//
+// closeInfo is a map, indexed by connection ID, to the individual
+// sessions IDs that those connections belong to. The values are
+// ignored; the parameter is expected as such just for convenience of
+// its caller.
+func (w *Worker) makeCloseConnectionRequest(closeInfo map[string]string) *pbs.CloseConnectionRequest {
+	closeData := make([]*pbs.CloseConnectionRequestData, 0, len(closeInfo))
+	for connId := range closeInfo {
 		closeData = append(closeData, &pbs.CloseConnectionRequestData{
 			ConnectionId: connId,
 			Reason:       session.UnknownReason.String(),
 		})
 	}
-	closeInfo := &pbs.CloseConnectionRequest{
+
+	return &pbs.CloseConnectionRequest{
 		CloseRequestData: closeData,
 	}
+}
 
-	connStatus, err := w.closeConnection(ctx, closeInfo)
-	if err != nil {
-		return err
+// makeSessionCloseInfo takes the response from CloseConnections and
+// our original closeInfo map and makes a map of slices, indexed by
+// session ID, of all of the connection responses. This allows us to
+// easily lock on session once for all connections in
+// setCloseTimeForResponse.
+func (w *Worker) makeSessionCloseInfo(
+	closeInfo map[string]string,
+	response *pbs.CloseConnectionResponse,
+) map[string][]*pbs.CloseConnectionResponseData {
+	if closeInfo == nil {
+		// Should never happen, panic if it does. Results will be
+		// undefined.
+		panic(errMakeSessionCloseInfoNilCloseInfo)
 	}
-	closedIds := make([]string, 0, len(connStatus.GetCloseResponseData()))
 
-	// Here we build a reverse map from closeMap, that is, session ID to
-	// connection IDs, for more efficient locking
-	revMap := make(map[string][]*pbs.CloseConnectionResponseData)
-	for _, v := range connStatus.GetCloseResponseData() {
-		revMap[closeMap[v.GetConnectionId()]] = append(revMap[closeMap[v.GetConnectionId()]], v)
+	result := make(map[string][]*pbs.CloseConnectionResponseData)
+	for _, v := range response.GetCloseResponseData() {
+		result[closeInfo[v.GetConnectionId()]] = append(result[closeInfo[v.GetConnectionId()]], v)
 	}
-	for k, v := range revMap {
-		siRaw, ok := w.sessionInfoMap.Load(k)
+
+	return result
+}
+
+// setCloseTimeForResponse iterates a CloseConnectionResponse and
+// sets the close time for any connection found to be closed to the
+// current time.
+//
+// sessionCloseInfo can be derived from the closeInfo supplied to
+// makeCloseConnectionRequest through reverseCloseInfo, which creates
+// a session ID to connection ID mapping.
+//
+// A non-zero error count does not necessarily mean the operation
+// failed, as some connections may have been marked as closed. The
+// actual list of connection IDs closed is returned as the first
+// return value.
+func (w *Worker) setCloseTimeForResponse(sessionCloseInfo map[string][]*pbs.CloseConnectionResponseData) ([]string, []error) {
+	closedIds := make([]string, 0)
+	var errors []error
+	for sessionId, responses := range sessionCloseInfo {
+		siRaw, ok := w.sessionInfoMap.Load(sessionId)
 		if !ok {
-			w.logger.Warn("could not find session ID in info map after closing connections", "session_id", k)
+			errors = append(errors, fmt.Errorf("could not find session ID %q in local state after closing connections", sessionId))
 			continue
 		}
+
 		si := siRaw.(*sessionInfo)
 		si.Lock()
-		for _, connResult := range v {
-			ci := si.connInfoMap[connResult.GetConnectionId()]
-			ci.status = connResult.GetStatus()
+
+		for _, response := range responses {
+			ci, ok := si.connInfoMap[response.GetConnectionId()]
+			if !ok {
+				errors = append(errors,
+					fmt.Errorf(
+						"could not find connection ID %q for session ID %q in local state after closing connections",
+						response.GetConnectionId(),
+						sessionId,
+					),
+				)
+				continue
+			}
+
+			ci.status = response.GetStatus()
 			if ci.status == pbs.CONNECTIONSTATUS_CONNECTIONSTATUS_CLOSED {
 				ci.closeTime = time.Now()
+				closedIds = append(closedIds, ci.id)
 			}
 		}
+
 		si.Unlock()
 	}
-	w.logger.Trace("connections successfully marked closed", "connection_ids", closedIds)
-	return nil
+
+	return closedIds, errors
 }