+++ internal/servers/worker/worker.go
@@ -28,6 +28,7 @@ type Worker struct {
 	started     *ua.Bool
 
 	controllerStatusConn *atomic.Value
+	workerStartTime      time.Time
 	lastStatusSuccess    *atomic.Value
 
 	controllerResolver *atomic.Value
@@ -60,6 +61,10 @@ func New(conf *Config) (*Worker, error) {
 	w.lastStatusSuccess.Store((*LastStatusInformation)(nil))
 	w.controllerResolver.Store((*manual.Resolver)(nil))
 
+	if conf.RawConfig.Worker == nil {
+		conf.RawConfig.Worker = new(config.Worker)
+	}
+
 	w.ParseAndStoreTags(conf.RawConfig.Worker.Tags)
 
 	if conf.SecureRandomReader == nil {
@@ -67,9 +72,6 @@ func New(conf *Config) (*Worker, error) {
 	}
 
 	var err error
-	if conf.RawConfig.Worker == nil {
-		conf.RawConfig.Worker = new(config.Worker)
-	}
 	if conf.RawConfig.Worker.Name == "" {
 		if conf.RawConfig.Worker.Name, err = base62.Random(10); err != nil {
 			return nil, fmt.Errorf("error auto-generating worker name: %w", err)
@@ -115,6 +117,7 @@ func (w *Worker) Start() error {
 	}
 
 	w.startStatusTicking(w.baseContext)
+	w.workerStartTime = time.Now()
 	w.started.Store(true)
 
 	return nil